feat(kp-cloud): 修改浮点数据类型

+++ kp-cloud/kp-system/src/main/java/org/dromara/kp/system/domain/KpChargeAccount.java
@@ -42,17 +42,17 @@ public class KpChargeAccount extends TenantEntity {
     /**
      * 性别
      */
-    private Long sex;
+    private Integer sex;
 
     /**
      * 账户类型
      */
-    private Long accoutType;
+    private Integer accoutType;
 
     /**
      * 禁用
      */
-    private Long disableFlag;
+    private Integer disableFlag;
 
     /**
      * 注册时间
